Universal functions: getRandomPositiveInteger, checkStringLength,... moved to 'util.js'. Data generation realization moved to 'data.js'.
